Execute some commands asynchronously
VTR fundamentally works by sending commands to tmux via the `tmux`
executable, and it order to do this it has to shell out. This can be
costly depending on how long it takes for your shell to boot. In an
effort to decrease this time, this commit makes it possible for some
commands (like the main command being run) to run asynchronously. This
functionality is provided by Neovim's job control feature, so obviously
this only works if you're using Neovim; otherwise the standard `system`
function built into Vim is used. (Unfortunately, not *all* of the
commands can be run asynchronously -- we need to get information from
tmux about the current runner pane before we can run the command in that
pane, for instance.)